Here is the generated commit message:
Commit has to be a summary in just one line.
New text file created.

Detailed description
This change adds a new text file, which can be used for testing or as a placeholder.
The file contains some sample text that can be modified as needed.
This change helps with testing and allows for more flexibility in the future.

Questions:
1. What changes were made?
    A new text file was created.
2. Why were these changes made?
    This change makes it easier to test and modify the project.
3. How do these changes address the problem?
    The change addresses the issue of not having a file for testing purposes.
4. Are there any side effects?
    No, there are no known side effects.

Change Summary:
New text file created.